+++ database/migrations/2025_10_21_000004_create_service_specifications_table.php
@@ -9,14 +9,16 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('service_specifications', function (Blueprint $table) {
-            $table->uuid('id')->default(DB::raw('(UUID())'))->primary();
-            $table->foreignUuid('service_id')->constrained()->onDelete('cascade');
-            $table->enum('type', ['principal', 'general']); // Para diferenciar
-            $table->string('title'); // Título de la especificación
-            $table->text('description'); // Descripción detallada
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('service_specifications')) {
+            Schema::create('service_specifications', function (Blueprint $table) {
+                $table->uuid('id')->default(DB::raw('(UUID())'))->primary();
+                $table->foreignUuid('service_id')->constrained()->onDelete('cascade');
+                $table->enum('type', ['principal', 'general']); // Para diferenciar
+                $table->string('title'); // Título de la especificación
+                $table->text('description'); // Descripción detallada
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
